+++ leda/components/Tooltip/index.tsx
@@ -35,7 +35,7 @@ export const Tooltip = React.forwardRef((props: TooltipProps, ref?: React.Ref<To
     tooltipRef,
   });
 
-  const tooltipClassNames = getClassNames(position ? theme[position] : theme.tooltip);
+  const tooltipClassNames = getClassNames(position ? `${theme.tooltip} ${position}` : theme.tooltip);
 
   // добавление обертки если нужно
   const element = React.isValidElement(children) ? children : (