+++ leda/components/Tooltip/theme.ts
@@ -5,6 +5,10 @@ export const defaultTooltipTheme = {
   left:                     'tooltip left',
   right:                    'tooltip right',
   bottom:                   'tooltip bottom',
+  topLeft:                  'tooltip top-left',
+  topRight:                 'tooltip top-right',
+  bottomLeft:               'tooltip bottom-left',
+  bottomRight:              'tooltip bottom-right',
   wrapper:                  'tooltip-wrapper',
 };
 /* eslint-enable key-spacing */