+++ leda/components/Tooltip/types.ts
@@ -3,6 +3,77 @@ import { COMPONENTS_NAMESPACES } from '~/constants';
 import { PartialGlobalDefaultTheme } from '~/utils/useTheme';
 
 /**
+ * Type of helper that checks corner position of tooltip
+ */
+export interface CheckCornerPosition {
+  (
+  /** Flag defines if tooltip is outside of right border of view port */
+    isOutsideRightBorder: boolean,
+  /** Flag defines if tooltip is outside of left border of view port */
+    isOutsideLeftBorder: boolean,
+  /** Flag defines if tooltip is outside of top border of view port */
+    isOutsideTopBorder: boolean,
+  /** Flag defines if tooltip is outside of bottom border of view port */
+    isOutsideBottomBorder: boolean,
+  ): {
+    /** Flag defines if tooltip is in one of bottom corners */
+    isBottomCorners: boolean,
+    /** Flag defines if tooltip is in one of top corners */
+    isTopCorners: boolean,
+    /** Flag defines if tooltip is in one of right corners */
+    isRightCorners: boolean,
+    /** Flag defines if tooltip is in one of left corners */
+    isLeftCorners: boolean,
+  },
+}
+
+/**
+ * Type of helper that checks horizontal position of tooltip
+ */
+export interface CheckHorizontalPosition {
+  (
+    /** DOMRect object of target element */
+    elementRect: DOMRect,
+    /** DOMRect object of tooltip */
+    tooltipRect: DOMRect,
+    /** Width or height of tooltip's arrow  */
+    arrowSize: number,
+  ): {
+    /** Flag defines if tooltip can be show on left */
+    isLeft: boolean,
+    /** Flag defines if tooltip can be show on right */
+    isRight: boolean,
+    /** Flag defines if tooltip is outside of top border of view port */
+    isOutsideTopBorder: boolean,
+    /** Flag defines if tooltip is outside of bottom border of view port */
+    isOutsideBottomBorder: boolean,
+  },
+}
+
+/**
+ * Type of helper that checks vertical position of tooltip
+ */
+export interface CheckVerticalPosition {
+  (
+    /** DOMRect object of target element */
+    elementRect: DOMRect,
+    /** DOMRect object of tooltip */
+    tooltipRect: DOMRect,
+    /** Width or height of tooltip's arrow  */
+    arrowSize: number,
+  ): {
+    /** Flag defines if tooltip can be show on top */
+    isTop: boolean,
+    /** Flag defines if tooltip can be show on bottom */
+    isBottom: boolean,
+    /** Flag defines if tooltip is outside of right border of view port */
+    isOutsideRightBorder: boolean,
+    /** Flag defines if tooltip is outside of left border of view port */
+    isOutsideLeftBorder: boolean,
+  },
+}
+
+/**
  * Type of helper that calculates offsets
  * based on tooltip position
  */
@@ -56,7 +127,7 @@ export interface TooltipBodyProps {
 /**
  * Position of tooltip against target element
  */
-export type TooltipPosition = 'top' | 'right' | 'bottom' | 'left' | undefined;
+export type TooltipPosition = 'top' | 'right' | 'bottom' | 'left' | 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right' | undefined;
 
 export interface TooltipProps {
   /** Размер стрелки тултипа в px */