+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - DatePicker, TimePicker: Fixed validation
 - MultiSelect: Input is cleared after selecting item
+- ToolTip: Added corner position
 - Validation: Fixed validation in controlled mode
 
 ## [0.33.0] - 2020-10-26